ensure the build fails when more than one source file is missing
The 'analyze' phase of the build process completes when all source files are analyzed (scanned). A source file is analized when all its dependencies are resolved or rejected. When the sources stream is complete, _all_ pending source file promises need to be rejected--not only the first one. If some of the deferred dependencies remain pending, the analysis of the toplevel source file never continues after 'await this._cache.dependencyGraph.whenReady(resolvedUrl);'. That leads of a premature end of the analysis phase (without any errors).